Cherry-pick #3063 to dfinity/v0.26
This patch is taken from
https://github.com/bytecodealliance/wasmtime/pull/3079
which in turn a cherry-pick of #3063 adjusted for v0.26:
https://github.com/bytecodealliance/wasmtime/pull/3063

- Restore POSIX signal handling on MacOS behind a feature flag

As described in Issue #3052, the switch to Mach Exception handling
removed unix::StoreExt from the public API of crate on MacOS.
That is a breaking change and makes it difficult for some
application to upgrade to the current stable Wasmtime.

As a workaround this PR introduces a feature flag called
posix-signals-on-macos that restores the old behaviour on MacOS.
The flag is disabled by default.

- Fix test guard

- Fix formatting in the test