Properly update detected ISA features in `builder_with_options()`
Currently `builder_with_options()` enables a feature if it is
detected as supported at runtime. However the function does not
disable unsupported features. This is problematic if a feature
is set to true by default, which is the case for SSE features
(see `cranelift_codegen_meta::isa::x86`).

On CPU that do not support SSE 4.1 this causes an error:
`compilation settings are not compatible with the native host`.

That is because:
- `Engine` is constructed with SSE features enabled by default,
- `Module` performs runtime detection of features and fails.

This PR changes `builder_with_options()` to clear unsupported
features.